Implement comment format cop
If someone writes a comment under a TODO but does not indent it, the
comment will not be listed and then the todo is empty and doesn't make
sense.

This cop notifies the user if the todo format is incorrect. In addition
I've implemented an autocorrect which I ran on core and fixed over 800
incorrectly formatted todo comments.